requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
aiosqlite==0.20.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comtypes==1.4.12
contourpy==1.3.2
cycler==0.12.1
discord.py==2.6.2
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.10
fonttools==4.59.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.1
magic-filter==1.0.12
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mss==10.1.0
multidict==6.6.4
noise==1.2.2
numpy==2.2.6
opencv-python==4.12.0.88
orjson==3.11.3
packaging==25.0
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pefile==2023.2.7
pillow==10.4.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.1
PyQt6_sip==13.10.2
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyTelegramBotAPI==4.29.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==20.6
pyttsx3==2.99
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.0
rsa==4.9.1
screeninfo==0.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
srt==3.5.3
starlette==0.37.2
tqdm==4.67.1
ttkbootstrap==1.14.2
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
ursina==7.0.0
uvicorn==0.30.1
vosk==0.3.45
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
